feat(trip): NewTripModal → /trips/new page (modal-to-fullpage migration v2.19.5)
DESIGN.md 2026-05-03「Modal vs Full Page」規則第二個遷移 PR (PR series 3/4)：
NewTripModal → NewTripPage @ /trips/new。同 pattern 套用 PR #428
(EditTripModal → /trip/:id/edit)。

What changed:
- NEW src/pages/NewTripPage.tsx (~860 LOC) — AppShell + sticky TitleBar
  (返回 + 「新增行程」+ 建立 action with .tp-titlebar-action.is-primary) +
  content (form max-width 720) + bottom sticky actions row。Form 邏輯 / state /
  handlers 完全沿用原 NewTripModal v2.19.0 (含 multi-dest sortable + day
  quota stepper + flex date stepper + month carousel + popular/recent dest
  chips + POI search debounce)。
- DEL src/components/trip/NewTripModal.tsx (~1028 LOC)
- Route: /trips/new mounted top-level (not under /trip/:tripId/* parent，
  因為這是建新行程，沒有現存 tripId)
- src/contexts/NewTripContext.tsx 簡化：不再 mount modal、無 modal state +
  handler，改 openModal() = navigate('/trips/new')。API 名稱保留相容性給
  既有 4 個 caller (DesktopSidebar via context、TripsListPage 卡片 + hero、
  GlobalMapPage)，行為對 caller 透明。
- main.tsx 加 lazy import + route entry。

Tests:
- DEL tests/unit/new-trip-modal.test.tsx + new-trip-modal-multidest.test.tsx
  (原本 render NewTripModal 直接驗 portal — page 版需要 router context，
  重寫成本 > 重新覆蓋價值，留 follow-up 補 NewTripPage 等價測試)
- UPDATE tests/unit/trips-list-page.test.tsx 2 個 assertion (modal 不再 mount
  on click)
- UPDATE tests/unit/mockup-typography-compliance.test.ts (grep 來源 path 改
  NewTripModal.tsx → NewTripPage.tsx)

Verification: 1332 unit + 603 API tests pass。typecheck (src/ + functions/) clean。

剩 PR 4: AddStopModal → /trip/:id/add-stop。Audit 還補發現:
- AccountPage 「登出」 inline modal (P3，confirm-style 應改 ConfirmModal)
- ConnectedAppsPage 「撤銷授權」 inline modal (P3，同上)
- DeveloperAppsPage create-app modal (P1，9+ 欄位，必走全頁)
- DeveloperAppsPage secret-display modal (P3，confirm-style)
- ExplorePage trip-picker modal (P2，picker dialog)
詳見 DESIGN.md audit table 待 PR 補。
