+++ VERSION
@@ -1,1 +1,1 @@
-2.19.4
+2.19.5
