+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tripline",
-  "version": "2.19.4",
+  "version": "2.19.5",
   "private": true,
   "scripts": {
     "dev": "concurrently -n vite,api -c cyan,green \"vite\" \"wrangler pages dev dist --local --port 8788\"",
